// src/filter/dtos/temp-zod-schemas.ts
export const WorkflowRuntimeDataWhereInputSchema = z.lazy(() =>
  z
    .object({
      id: zStringFilterStringUnion.optional(),
      endUserId: zStringNullableFilterStringNullUnion.optional(),
      businessId: zStringNullableFilterStringNullUnion.optional(),
      workflowDefinitionId: zStringFilterStringUnion.optional(),
      workflowDefinitionVersion: zStringFilterStringUnion.optional(),
      context: z.unknown().optional(),
      assigneeId: zStringNullableFilterStringNullUnion.optional(),
      config: z.unknown().optional(),
      state: zStringNullableFilterStringNullUnion.optional(),
      status: zStringFilterStringUnion.optional(),
      createdAt: zDateTimeFilterDateStringUnion.optional(),
      updatedAt: zDateTimeFilterDateStringUnion.optional(),
      createdBy: zStringFilterStringUnion.optional(),
      endUser: z.union([EndUserRelationFilterSchema, EndUserWhereInputSchema, z.null()]).optional(),
      business: z
        .union([BusinessRelationFilterSchema, BusinessWhereInputSchema, z.null()])
        .optional(),
      workflowDefinition: WorkflowDefinitionRelationFilterSchema.or(
        WorkflowDefinitionWhereInputSchema,
      ).optional(),
      assignee: UserWhereInputSchema.or(UserRelationFilterSchema).optional(),
    })
    .strict(),
);
// ...
export const WorkflowRuntimeDataSelectSchema = z.object({
  id: z.boolean().optional(),
  endUserId: z.boolean().optional(),
  businessId: z.boolean().optional(),
  assigneeId: z.boolean().optional(),
  workflowDefinitionId: z.boolean().optional(),
  workflowDefinitionVersion: z.boolean().optional(),
  context: z.boolean().optional(),
  state: z.boolean().optional(),
  status: z.boolean().optional(),
  createdAt: z.boolean().optional(),
  updatedAt: z.boolean().optional(),
  createdBy: z.boolean().optional(),
  resolvedAt: z.boolean().optional(),
  endUser: z
    .union([
      z.boolean(),
      z.object({
        select: z
          .lazy(() => EndUserSelectSchema.strict())
          .nullable()
          .optional(),
      }),
    ])
    .optional(),
  business: z
    .union([
      z.boolean(),
      z.object({
        select: z
          .lazy(() => BusinessSelectSchema.strict())
          .nullable()
          .optional(),
      }),
    ])
    .optional(),
  workflowDefinition: z
    .union([
      z.boolean(),
      z.object({
        select: z
          .lazy(() => WorkflowDefinitionSelectSchema.strict())
          .nullable()
          .optional(),
      }),
    ])
    .optional(),
  assignee: z
    .union([
      z.boolean(),
      z.object({
        select: z
          .lazy(() => UserSelectSchema.strict())
          .nullable()
          .optional(),
      }),
    ])
    .optional(),
});
// ...
const WorkflowRuntimeDataFilterQuerySchema = z
  .object({
    select: WorkflowRuntimeDataSelectSchema.strict()
      .refine(
        (v: Record<PropertyKey, unknown>) => Object.keys(v).length > 0,
        'At least one `select` field must be provided',
      )
      .optional(),
    where: WorkflowRuntimeDataWhereInputSchema.optional(),
  })
  .refine(v => v.select || v.where, 'At least `query.select` or `query.where` must be provided');

export const FilterCreateSchema = z.object({
  name: z.string(),
  entity: z.enum(['individuals', 'businesses']),
  query: WorkflowRuntimeDataFilterQuerySchema,
});